refactor: Update imports after renaming main to core
